refactor: 페이지와 컴포넌트 폴더 분리
