+++ week2/Assignment2/src/components/mainStatic.js
