+++ week2/Assignment2/src/components/renderFunction.js
